Move test from dualsimulation to its parent
